change logging global var to use class
