+++ src/org/dsty/logging/package-info.groovy
@@ -1,0 +1,4 @@
+/**
+ * Logging and console output related functionality.
+ */
+package org.dsty.logging
